// Admin/src/models/CategoryList.Model.ts
export interface ICategoryModel {
    Active: boolean
    Created: string
    CreatedBy: number
    Description: null
    DisplayInList: boolean
    DisplayName: string
    EntityTypeId: number
    EnumName: string
    Id: number
    IsValid: boolean
    Modified: string
    ModifiedBy: number
    Name: string
    RowOrder: number
    ValidationErrors: {}
}